gitlab-ci: disable env var inheritance to downstream pipeline
By default, all downstream triggered pipelines inherit all of the
global env vars from the upstream project.  This was resulting
in the downstream pipeline (unexpectedly) inheriting the setting
for the HT_SWDEV_DOCKER_IMAGE env var which had an older version
of the docker container.  All triggered pipelines were thus using
the wrong/old version of the docker container and failing to build.