sn-stack: dpdk service may use locally built container for dpdk applications
Using a locally built DPDK container is necessary when running a CI
artifact (built on Intel CPUs) on a development server which has AMD
CPUs.  The Intel-built DPDK code includes opcodes which are not
present on AMD CPUs and thus results in a core dump.

This provides a hook to allow the 'smartnic-dpdk' container to provide
a locally built DPDK environment.