build: Fix use of PWD outside of RUN
Since PWD isn't passed in as a docker build argument from build.sh, it expands
to empty when used with the COPY instruction for copying the entire source tree.

This is seen in the build log for the package CI job:
----------CUT----------
11 [stage-0 3/7] COPY  /sn-fw/source
11 sha256:757582f3e1ce21256d32db2a332b922e511b409d826d5e770b742181f9335c77
11 DONE 0.2s
----------CUT----------
Found at https://gitlab.es.net/ht/esnet-smartnic-fw/-/jobs/232052#L1133

While this appears to be functioning with the Docker build tools in
docker:20.10.21-dind, it fails with the newer 24.0.5 version used in
Ubuntu 22.04.