test: Remove dependencies of smartnic-fw-test service from docker compose
The single Hello World test currently has no depedencies on other services. Any
dependencies new tests may have should be made explicit at the time of creation.

The docker compose COMPOSE_PROFILES environment variable should be used to
select the minimal set of other services to start prior to executing the test
service, as the exact set may change depending on the suites being run.

Test suites that depend on other services will need to define a setup fixture
to ensure that the environment is prepared prior to execution. This allows for
each fixture to customize the criteria for health and the methods used to verify
it in the needed services.