test: Disable writing Python byte code into mounted test library
The test engine is currently running as the root user within the container,
leading the compiled byte code .pyc files generated by Python to be written
back to the sn-stack/test mount point as root on the host.

For now, disable writing the byte code and mount sn-stack/test as read-only.
This does not affect test result output files written to sn-stack/scratch.