libvitisnetp4drv: Merge Vitis driver static libraries into a single one
This change merges multiple sdnet_X header and static library files into
a single one. Header files from each sdnet_X/include directory are simply
copied into the same location. The object files from each sdnet_X/lib
archive is extracted on top of each other into a staging directory and
then a new libvitisnetp4drv.a is created containing the merged result.
Duplicate header and object files are overwritten by following copies.

This strategy is based on the observation that the Vitis driver only seems
to be generated with a single .h/.c pair. As a result, only a single object
file per archive has any symbols containing the sdnet_X string. All other
object files appear to be independent of the the sdnet_X string, with the
only differences being in compile paths and timestamps.

The above assumptions need to be verified with Xilinx.