sn-p4-agent: Allocate struct snp4_info_pipeline on heap due to stack size limit
A single instance of the struct snp4_info_pipeline is 4360200 bytes. Adding a
second leads to a segfault when instantiating a SmartnicP4Impl on the stack of
RunServer (due to the default stack size being 8192K bytes).

Changed cached pipeline info to an array of pointers to be allocated in the
constructor and added a destructor for cleaning up afterwards.