sn-cfg: Track agent device objects by reference
The std::vector template class gives read-only references to the members by
means of creating a copy of the data. To make the vector's members writeable,
it's necessary to push a reference to an existing object.

This change updates the vector of managed devices to use references in order to
permit modification in subsequent init stages after the device vector has been
populated.