sn-p4: Wrap verbose logging in a debug flag
The verbose logging has been found to reduce server operation time. All
debug logs are now wrapped by a flag to silence them for normal deployment.

The debug flag has been made a static parameter, requiring a re-compile to
enable. It will be added as a command line parameter in the future.