sn-cfg/sn-p4: Allow passing the TLS root certs as a raw byte string
This permits users of the connect_client function to provide the TLS root
certificate chain directly as a bytes object. This feature is primarily
intended to support communicating with a server using a self-signed
certificate (since the gRPC Python API doesn't expose the options for how
the peer verification is performed).