sn-cfg: Don't display module baud rate of 0
According to the standard, a baud rate of 0 means the following:
--------------------------------------------------------------------------------
Section 6.3.6 (SFF-8636 rev 2.11, page 50) [upper page 0, register 140]:
A value of 0 indicates the baud rate is not specified and must be determined
from the module technology.
--------------------------------------------------------------------------------

The "module technology" is encoded in the spec compliance codes listed under the
"Device" section of the CLI output.