protobuf: Update Python poetry tool usage to align with 2.0.0 updates to include keyword
According to the release notes for poetry 2.0.0 (released on 2025-01-05), the
default behaviour of the "include" keyword's short form in the TOML file has
changed to apply only to an sdist build. The prior behaviour of applying to
both sdist and wheel builds can be achieved by using the full syntax and
explicitly specyfing both formats.

Refer to:
- https://python-poetry.org/blog/announcing-poetry-2.0.0#consistent-include-behavior
- https://python-poetry.org/docs/pyproject/#exclude-and-include