build: Add support for attaching optional metadata to vitisnetp4 driver
When a metadata.yaml file is included in the HW artifacts for a pipeline's
vitisnetp4 driver directory, it's contents will be translated into C and
made available to FW via the interface structure.

The only metadata currently supported is a mapping to provide name aliases
for the indices of a P4 counter block. The format is:
--- START metadata.yaml ---
counter_blocks:
  <block-name0>:
    counters:
      0:
        alias: <alias-name0>
      1:
        alias: <alias-name1>
      ...

  <block-name1>:
    ...
--- END metadata.yaml ---

An example of providing name aliases for 2 counter blocks containg 4 counters
each:
--- START metadata.yaml ---
counter_blocks:
  packet_rx_counter:
    counters:
      0:
        alias: packet_rx_counter_ok
      1:
        alias: packet_rx_counter_non_eth
      2:
        alias: packet_rx_counter_non_ip
      3:
        alias: packet_rx_counter_non_tcpudp

  notrack_result_counter:
    counters:
      0:
        alias: notrack_result_counter_track_both
      1:
        alias: notrack_result_counter_flow_only
      2:
        alias: notrack_result_counter_non_ip
      3:
        alias: notrack_result_counter_non_tcpudp
--- END metadata.yaml ---

wip-build-metadata