stack: Remove unused hostname from test container to allow use of conflicting network_mode
Since the smartnic-fw-test container is intended for running Robot Framework tests, it should
only act as a client of other services and is not expected to ever need to offer services (if
so, a separate container should be created for this new service). As such, the use of the
hostname attribute in the FW docker-compose.yml prevents overrides in stacked compose files
from setting the network_mode attribute to bind the smartnic-fw-test container to another
container’s network namespace (this is needed to allow sharing virtual network devices created
in a container with tests running in smartnic-fw-test).